:shirt: Rename methods folder to controllers, update require statements.
